Turn on optimisation for the client and server. Server apparently freezes after it has been running with two clients for, say, 5-10 minutes.
Randomise the various delays. Seems to have reduced the number of errors. Between this change and the change in onPairing in server.cpp (to a wait without a timeout), pairing now seems to work without a glitch. 

The WORK message is repeatedly broadcasted by the server for about 1 second. In addition, clients wait from 1s to 1.5s after receiving the WORK message. This seems to have healed the issue with some clients not receiving WORK.

Improve the way status is printed so it's shown at least once per state change.

Force rebuilds when ./deploy is called. Use cat instead of screen in ./monitor because it's more robust.